+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - rendering backstory for Cthulhu characters
 - managing equipment for Cthulhu characters
+- importing D&D characters from D&D Beyond
 
 ## [0.4.35] - 2026-06-13
 ### Added